postgresql-watcher/requirements.txt
ops>=2.0.0
psycopg2-binary>=2.9.0
pysyncobj>=0.3.0
